requirements.txt
python-telegram-bot==20.3
openai==1.16.1
python-dotenv==1.0.1
pyttsx3==2.90
SpeechRecognition==3.10
pydub==0.25.1